% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_flow_events.R
\name{compile_flow_events}
\alias{compile_flow_events}
\title{Compile Flow Cytometry Events and Polyfunctionality Statistics}
\usage{
compile_flow_events(
  gs,
  output_nodes,
  parent_node,
  cytokine_nodes,
  pData_cols = c("BATCH", "PTID", "STIM", "VISITNO", "Run Num", "Collection Num",
    "Replicate", "SAMP_ORD"),
  do.comp = FALSE,
  do.biexp = FALSE,
  do.asinh = TRUE,
  do.asym = TRUE,
  asym_root = 2,
  cofactor = 500,
  stim_to_exclude = NULL
)
}
\arguments{
\item{gs}{A \code{GatingSet} object.}

\item{output_nodes}{Character vector. Names of the FlowJo gates to extract boolean positivity (e.g., "IFNg+").}

\item{parent_node}{Character. The parent gate to filter cells (e.g., "CD4"). Only cells in this gate are returned.}

\item{cytokine_nodes}{Character vector. Subset of \code{output_nodes} used to count positive cytokines per cell.}

\item{pData_cols}{Character vector. Columns from \code{pData(gs)} to include in the output.
Default includes BATCH, PTID, STIM, VISITNO, Run Num, Replicate, SAMP_ORD}

\item{do.comp}{Logical. If \code{TRUE}, extracts compensated raw values.}

\item{do.biexp}{Logical. If \code{TRUE}, extracts FlowJo's biexponential transformed values.}

\item{do.asinh}{Logical. If \code{TRUE}, applies Arcsinh transformation using \code{do_asinh_local}.}

\item{do.asym}{Logical. If \code{TRUE}, applies an additional asymmetric root transformation to Arcsinh values.}

\item{asym_root}{Numeric. The root used for asymmetric transformation. Default is 2.}

\item{cofactor}{Numeric. Co-factor for Arcsinh transformation. Default is 500.}

\item{stim_to_exclude}{Character vector. List of stimulations to filter out from the final results.}
}
\value{
A \code{list} containing two \code{data.table} elements:
\itemize{
  \item \code{exprs}: Single-cell level intensities and gate positivity.
  \item \code{cytnum}: Summarized counts of cytokine-positive vs. negative cells per sample.
}
}
\description{
This wrapper function serves as the primary entry point for data extraction in the \code{ICSR}
pipeline. It simultaneously extracts detailed single-cell expression data and
summarized cytokine polyfunctionality counts by calling \code{extract_flow_exprs_data}
and \code{extract_CYTNUM_data}.
}
