% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gating_set.R
\name{create_gating_set}
\alias{create_gating_set}
\title{create_gating_set}
\usage{
create_gating_set(
  assayid,
  xml_path,
  fcs_path,
  batchData,
  sample_name = "Samples",
  xml_keywords = c("$FIL", "$TOT", "EXPERIMENT NAME", "Sample Order", "Stim",
    "Replicate"),
  output_path = NULL
)
}
\arguments{
\item{assayid}{`numeric` or `character`. batch ID: used for merging Master thaw list and `pData()`.}

\item{xml_path}{xml path (FlowJo .xml).}

\item{fcs_path}{fcs path.}

\item{batchData}{Master thaw list. Need BATCH, SAMP_ORD, PTID, VISITNO and RUNNUM.}

\item{sample_name}{`numeric` or `character`. The name or index of the group of samples to be imported.}

\item{xml_keywords}{`character` vector specifying the keywords to be extracted as pData of GatingSet.}

\item{output_path}{Output path for the new GatingSet object.}
}
\value{
GatingSet R obj.
}
\description{
create_gating_set
}
