% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_filtering.R
\name{do_filtering}
\alias{do_filtering}
\title{Filtering}
\usage{
do_filtering(
  dt,
  remove_dup = FALSE,
  dt.MTL_dup = NULL,
  remove_participants = FALSE,
  participants_list = NULL,
  remove_nsub = FALSE,
  cutoff_nsub = NULL,
  flag_unreliable = FALSE,
  unreliable_list_all = NULL,
  unreliable_list_by_stim = NULL
)
}
\arguments{
\item{dt}{input data.table. Including pData and fluorescence intensities (FI).}

\item{remove_dup}{TRUE or FALSE. If TRUE, remove duplicates (based on RUNNUM).}

\item{dt.MTL_dup}{Master thaw list to remove duplicates (based on RUNNUM). Need BATCH, SAMP_ORD, PTID, VISITNO and RUNNUM.}

\item{remove_participants}{TRUE or FALSE. If TRUE, remove participants, e.g. control participants.}

\item{participants_list}{Vector of removed participants.}

\item{remove_nsub}{TRUE or FALSE. If TRUE, remove samples based on total number of cells. Ex: 10,000 CD4+ T.}

\item{cutoff_nsub}{Cut-off for `remove_nsub`. Ex: 10,000 CD4+ T.}

\item{flag_unreliable}{TRUE or FALSE. Flag unreliable samples. If TRUE, need unreliable lists.}

\item{unreliable_list_all}{Vector of unreliable samples. Unreliable sample level must be BATCH_PTID_VISITNO_RUNNUM. Use with caution.}

\item{unreliable_list_by_stim}{Vector of unreliable samples. Unreliable sample level must be BATCH_PTID_VISITNO_RUNNUM_REPLICATE_STIM. Use with caution.}
}
\value{
data.table with pData, FlowJo positivity (positivity for each marker using the marker gates determined by the lab (FlowJo gating)), and FI after filtering.
}
\description{
Filtering
}
