% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot_annotation.R
\name{dotplot_annotation}
\alias{dotplot_annotation}
\title{dotplot_annotation}
\usage{
dotplot_annotation(
  dt,
  markers,
  flowjo_gates,
  cluster_col = "Leiden",
  level_marker = NULL,
  order = NULL
)
}
\arguments{
\item{dt}{data.table Input sample.}

\item{markers}{List of markers used for MFI calculation.}

\item{flowjo_gates}{flowJo gates.}

\item{cluster_col}{Cluster column.}

\item{level_marker}{Marker order to follow.}

\item{order}{Cluster order to follow.}
}
\value{
UMAP embeddings
}
\description{
dotplot_annotation
}
