% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_functions.R
\name{dt.ggplot_boolean}
\alias{dt.ggplot_boolean}
\title{dt.ggplot_boolean}
\usage{
dt.ggplot_boolean(dt, flowjo_gates, cluster_col = "Leiden")
}
\arguments{
\item{dt}{data.table Input sample.}

\item{flowjo_gates}{flowJo gates.}

\item{cluster_col}{Cluster column.}
}
\value{
data.table with marker positivity, gated
}
\description{
dt.ggplot_boolean
}
