% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_functions.R
\name{dt.ggplot_MFI}
\alias{dt.ggplot_MFI}
\title{dt.ggplot_MFI}
\usage{
dt.ggplot_MFI(dt, markers, cluster_col = "Leiden")
}
\arguments{
\item{dt}{data.table Input sample.}

\item{markers}{List of markers used for MFI calculation.}

\item{cluster_col}{Cluster column.}
}
\value{
data.table with MFI
}
\description{
dt.ggplot_MFI
}
