% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_functions.R
\name{x_axis}
\alias{x_axis}
\title{x_axis}
\usage{
x_axis(dt.boolean, order = NULL)
}
\arguments{
\item{dt.boolean}{data.table boolean input.}

\item{order}{cluster order to follow. By default, NULL.}
}
\value{
Vector for x axis corresponding to Flowjo based marker annotation.
}
\description{
x_axis
}
