% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_flow_events.R
\name{compile_flow_events}
\alias{compile_flow_events}
\title{compile_flow_events}
\usage{
compile_flow_events(
  gs,
  output_nodes,
  parent_node,
  cytokine_nodes,
  do.comp = FALSE,
  do.biexp = FALSE,
  do.asinh = TRUE,
  do.asym = TRUE,
  asym_root = 2,
  cofactor = 500,
  stim_to_exclude = NULL
)
}
\arguments{
\item{gs}{GatingSet Object.}

\item{output_nodes}{Flowjo gates.}

\item{parent_node}{Parent Flowjo gate.}

\item{cytokine_nodes}{Cytokine Flowjo gates.}

\item{do.comp}{TRUE or FALSE. If TRUE, get compensated values.}

\item{do.biexp}{TRUE or FALSE. If TRUE, get biexp transformed values using FlowJo transformation parameters.}

\item{do.asinh}{TRUE or FALSE. If TRUE, get arcsinh + asym transformed values.}

\item{asym_root}{Root used for asym transformation. By default 2.}

\item{cofactor}{Co-factor used for arcsinh transformation.}

\item{stim_to_exclude}{Stimulation(s) to be removed. By default NULL.}
}
\value{
list with two data.table elements 'exprs' and 'cytnum'
}
\description{
compile_flow_events
}
