% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report_QC_ICS.R
\name{create_report_QC_ICS}
\alias{create_report_QC_ICS}
\title{ICS QC Report}
\usage{
create_report_QC_ICS(
  dt.exprs,
  dt.cytnum,
  cytokine_nodes = NULL,
  markers = NULL,
  output_format = html_document(toc = TRUE, toc_depth = 6, theme = "yeti"),
  output_file = "ICS_QC_report.html",
  output_dir = getwd(),
  report_title = "ICS QC Report",
  report_author = NULL
)
}
\arguments{
\item{dt.exprs}{input data.table. Including FI.}

\item{dt.cytnum}{input data.table. Including # of cytokines per sample.}

\item{cytokine_nodes}{Cytokine Flowjo gates. By default, NULL.}

\item{markers}{List of markers (ridge plots). By default, NULL.}

\item{output_format}{output format in \link{render}. Default is \code{html_document(toc = TRUE, toc_depth = 6, theme = "yeti")}.}

\item{output_file}{output file name in \link{render}. Default is "ICS_QC_report.html".}

\item{output_dir}{output directory for report in \link{render}. Default is user's current directory.}

\item{report_title}{report title. Default is "ICS QC Report".}

\item{report_author}{report authors. Default is NULL.}
}
\description{
This function creates a data QC report.
}
