% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_asinh_local.R
\name{do_asinh_local}
\alias{do_asinh_local}
\title{arcsinh transformation (from Spectre R package)}
\usage{
do_asinh_local(
  dat,
  use.cols,
  cofactor = 5,
  append.cf = FALSE,
  reduce.noise = FALSE
)
}
\arguments{
\item{dat}{data.table Input sample.}

\item{use.cols}{Vector of character column names – these columns will be transformed and added to the data.table as new columns.}

\item{cofactor}{Co-factor to use for arcsinh transformation.}

\item{append.cf}{Appends the co-factor used to the end of the name of the transformed columns.}

\item{reduce.noise}{This is an experimental calculation which should reduce noise from negative values. Use with caution.}
}
\value{
arcsinh transformed values
}
\description{
arcsinh transformation (from Spectre R package)
}
