% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_CYTNUM_data.R
\name{extract_CYTNUM_data}
\alias{extract_CYTNUM_data}
\title{extract_CYTNUM_data}
\usage{
extract_CYTNUM_data(gs, parent_node, cytokine_nodes, stim_to_exclude = NULL)
}
\arguments{
\item{gs}{GatingSet Object.}

\item{parent_node}{Parent Flowjo gate.}

\item{cytokine_nodes}{Cytokine Flowjo gates.}

\item{stim_to_exclude}{Stimulation(s) to be removed. By default NULL.}
}
\value{
data.table with pData and number of cytokines expressed (based on FlowJo gating positivity).
}
\description{
extract_CYTNUM_data
}
