% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_flow_exprs_data.R
\name{extract_flow_exprs_data}
\alias{extract_flow_exprs_data}
\title{extract_flow_exprs_data}
\usage{
extract_flow_exprs_data(
  gs,
  output_nodes,
  parent_node,
  cytokine_nodes,
  do.comp = FALSE,
  do.biexp = FALSE,
  do.asinh = TRUE,
  do.asym = TRUE,
  asym_root = 2,
  cofactor = 500,
  stim_to_exclude = NULL
)
}
\arguments{
\item{gs}{GatingSet Object.}

\item{output_nodes}{Flowjo gates.}

\item{parent_node}{Parent Flowjo gate.}

\item{cytokine_nodes}{Cytokine Flowjo gates.}

\item{do.comp}{TRUE or FALSE. If TRUE, get compensated values.}

\item{do.biexp}{TRUE or FALSE. If TRUE, get biexp transformed values using FlowJo transformation parameters.}

\item{do.asinh}{TRUE or FALSE. If TRUE, get arcsinh + asym transformed values.}

\item{asym_root}{Root used for asym transformation. By default 2.}

\item{cofactor}{Co-factor used for arcsinh transformation.}

\item{stim_to_exclude}{Stimulation(s) to be removed. By default NULL.}
}
\value{
data.table with pData, FlowJo gating positivity and FI.
}
\description{
extract_flow_exprs_data
}
