% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leiden_local.R
\name{leiden_local}
\alias{leiden_local}
\title{Leiden algorithm}
\usage{
leiden_local(data, markers, k = 30, res = 1, niter = 10, seed = 1234)
}
\arguments{
\item{markers}{Markers to use for clustering.}

\item{k}{Number of k neighbors.}

\item{res}{Resolution. A parameter controlling the coarseness of the clusters.}

\item{niter}{Number of iterations to run the Leiden algorithm.}

\item{seed}{Seed for the random number generator.}
}
\value{
partition of Leiden clusters
}
\description{
Leiden algorithm
}
