% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMIMOSA.R
\name{runMIMOSA}
\alias{runMIMOSA}
\title{# Originally created by Daryl Morris
PURPOSE: Apply a response call to ICS (intracellular cytokine staining) data
 as produced by the HVTN lab}
\usage{
runMIMOSA(
  INFILE = NULL,
  OUTFILE = NULL,
  CLUSTERS = paste("Leiden:", 1:10),
  MIMOSA_THRESHOLD_FDR = 0.01,
  FIT_METHOD = "mcmc",
  COMBINE = NULL,
  ANTIGENFILTER = NULL,
  MINSAMPLES = 10,
  ITER = 250000,
  BURN = 50000
)
}
\arguments{
\item{BURN}{}
}
\value{
.csv file with MIMOSA results
}
\description{
# Originally created by Daryl Morris
PURPOSE: Apply a response call to ICS (intracellular cytokine staining) data
 as produced by the HVTN lab
}
