% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAP_local.R
\name{UMAP_local}
\alias{UMAP_local}
\title{UMAP}
\usage{
UMAP_local(
  dt,
  markers,
  n_neighbors = 10,
  min_dist = 0.1,
  verbose = TRUE,
  n_components = 2,
  seed = 1234
)
}
\arguments{
\item{markers}{Markers to use for UMAP.}

\item{n_neighbors}{The size of local neighborhood. See `uwot::umap`.}

\item{min_dist}{The effective minimum distance between embedded points. See `uwot::umap`.}

\item{verbose}{If TRUE, log details to the console. See `uwot::umap`.}

\item{n_components}{The dimension of the space to embed into. See `uwot::umap`.}

\item{seed}{Seed for the random number generator.}
}
\value{
UMAP embeddings
}
\description{
UMAP
}
